Moving EmptyToNull and RedactionBehaviour to dev.guardrail.core
