+++ modules/core/src/main/scala/dev/guardrail/core/EmptyToNullBehaviour.scala
@@ -1,0 +1,5 @@
+package dev.guardrail.core
+
+sealed trait EmptyToNullBehaviour
+case object EmptyIsNull  extends EmptyToNullBehaviour
+case object EmptyIsEmpty extends EmptyToNullBehaviour
