+++ modules/core/src/main/scala/dev/guardrail/core/RedactionBehaviour.scala
@@ -1,0 +1,5 @@
+package dev.guardrail.core
+
+sealed trait RedactionBehaviour
+case object DataVisible  extends RedactionBehaviour
+case object DataRedacted extends RedactionBehaviour
