+++ modules/scala-http4s/src/main/scala/dev/guardrail/generators/Scala/Http4sServerGenerator.scala
@@ -6,7 +6,7 @@ import cats.implicits._
 import cats.Traverse
 import dev.guardrail.{ CustomExtractionField, RenderedRoutes, StrictProtocolElems, Target, TracingField, UserError }
 import dev.guardrail.core.{PathExtractor, Tracker}
-import dev.guardrail.extract.{ ServerRawResponse, TracingLabel }
+import dev.guardrail.core.extract.{ ServerRawResponse, TracingLabel }
 import dev.guardrail.generators.{ LanguageParameter, LanguageParameters }
 import dev.guardrail.generators.syntax._
 import dev.guardrail.generators.operations.TracingLabelFormatter
